Refactored naming of: - CoreData accessory class; - Datastore protocol.
