--- Tripper/Data/DAO/Protocols/DAO.swift
@@ -1,16 +1,0 @@
-//
-//  DAO Interface.swift
-//  Tripper
-//
-//  Created by Denis Cherniy on 11.02.2020.
-//  Copyright © 2020 Denis Cherniy. All rights reserved.
-//
-
-import Foundation
-
-protocol RoutePointDAO {
-    func fetchAll() -> [RoutePoint]
-    func insert(_ point: RoutePoint)
-    func update(_ point: RoutePoint)
-    func delete(_ point: RoutePoint)
-}
